requirements.txt
manim~=0.16.0
Pillow~9.2.0
